requirements.txt
arviz==0.20.0
biom-format==2.1.16
birdman==0.2.0
causal-learn==0.1.4.0
click==8.1.8
cmdstanpy==1.2.5
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
graphviz==0.20.3
h5netcdf==1.5.0
h5py==3.13.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
momentchi2==0.1.8
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
pydot==3.0.4
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
ruamel.yaml==0.18.10
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
stanio==0.5.1
statsmodels==0.14.4
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
wheel==0.45.1
xarray==2025.1.2
xarray-einstats==0.8.0
